+++ winterdrp/pipelines/winter/__init__.py
@@ -1,0 +1,1 @@
+from winterdrp.pipelines.winter.winter_pipeline import WINTERPipeline
