+++ winterdrp/pipelines/winter/constants.py
@@ -1,0 +1,1 @@
+winter_filters_map = {"Y": 1, "J": 2, "Hs": 3, "dark": 4}
